order add

+++ order/models.py
@@ -14,7 +14,7 @@ STATUS_CHOICES = (
 class OrderItem(models.Model):
     order = models.ForeignKey('Order', related_name='items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField(default=1)
+    quantity = models.PositiveSmallIntegerField(default=1)
 
 
 class Order(models.Model):
